+++ Task_1_1_1/app/build.gradle.kts
@@ -37,7 +37,7 @@ java {
 
 application {
     // Define the main class for the application.
-    mainClass.set("task_1_1_1.App")
+    mainClass.set("heapsort.App")
 }
 
 tasks.test {
