+++ Task_1_1_1/app/src/main/java/heapsort/App.java
@@ -1,7 +1,4 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package task_1_1_1;
+package heapsort;
 
 public class App {
     /**
