+++ Task_1_1_1/app/src/test/java/heapsort/AppTest.java
@@ -1,16 +1,15 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package task_1_1_1;
+package heapsort;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-
 import java.util.Arrays;
+import org.junit.jupiter.params.provider.CsvSource;
 
 public class AppTest {
     @Test
