+++ Task_1_1_1/build.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env sh
 
-javadoc -d docs -sourcepath ./app/src/main/java task_1_1_1
+javadoc -d docs -sourcepath ./app/src/main/java heapsort
 mkdir -p build
-javac ./app/src/main/java/task_1_1_1/App.java -d build
+javac ./app/src/main/java/heapsort/App.java -d build
 cd build
-java task_1_1_1.App
+java heapsort.App
